IP repo and sourcefiles for emulated DVS
Zip file contains three folders; DVS_IP is a repository which can be added directly to a Vivado Design. DVS is a HLS project which can be loaded. DVS_sourcefiles are accessed and edited by HLS. These are C files which can be opened in any code editor if required.